--- a/Robust.Server/GameStates/PvsData.cs
+++ b/Robust.Server/GameStates/PvsData.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using Robust.Shared.Collections;
 using Robust.Shared.GameObjects;
@@ -7,29 +8,31 @@ using Robust.Shared.GameStates;
 using Robust.Shared.Network;
 using Robust.Shared.Player;
 using Robust.Shared.Timing;
+using Robust.Shared.Utility;
 
 namespace Robust.Server.GameStates;
 
 /// <summary>
 /// Class for storing session specific PVS data.
 /// </summary>
-internal sealed class PvsSession(ICommonSession session)
+internal sealed class PvsSession(ICommonSession session, ResizableMemoryRegion<PvsData> memoryRegion)
 {
     public readonly ICommonSession Session = session;
-    public int Offset = -1;
+
+    public readonly ResizableMemoryRegion<PvsData> DataMemory = memoryRegion;
 
     public INetChannel Channel => Session.Channel;
 
     /// <summary>
-    /// All <see cref="EntityUid"/>s that this session saw during the last <see cref="PvsSystem.DirtyBufferSize"/> ticks.
+    /// All entities that this session saw during the last <see cref="PvsSystem.DirtyBufferSize"/> ticks.
     /// </summary>
-    public readonly OverflowDictionary<GameTick, List<IntPtr>> PreviouslySent = new(PvsSystem.DirtyBufferSize);
+    public readonly OverflowDictionary<GameTick, List<PvsIndex>> PreviouslySent = new(PvsSystem.DirtyBufferSize);
 
     /// <summary>
     /// <see cref="PreviouslySent"/> overflow in case a player's last ack is more than
     /// <see cref="PvsSystem.DirtyBufferSize"/> ticks behind the current tick.
     /// </summary>
-    public (GameTick Tick, List<IntPtr> SentEnts)? Overflow;
+    public (GameTick Tick, List<PvsIndex> SentEnts)? Overflow;
 
     /// <summary>
     /// The client's current visibility mask.
@@ -39,13 +42,13 @@ internal sealed class PvsSession(ICommonSession session)
     /// <summary>
     /// The list that is currently being prepared for sending.
     /// </summary>
-    public List<IntPtr>? ToSend;
+    public List<PvsIndex>? ToSend;
 
     /// <summary>
     /// The <see cref="ToSend"/> list from the previous tick. Also caches the current tick that the PVS leave message
     /// should belong to, in case the processing is ever run asynchronously with normal system/game ticking.
     /// </summary>
-    public (GameTick ToTick, List<IntPtr> PreviouslySent)? LastSent;
+    public (GameTick ToTick, List<PvsIndex> PreviouslySent)? LastSent;
 
     /// <summary>
     /// Visible chunks, sorted by proximity to the clients's viewers;
@@ -122,7 +125,10 @@ internal sealed class PvsSession(ICommonSession session)
 /// <summary>
 /// Class for storing session-specific information about when an entity was last sent to a player.
 /// </summary>
-[StructLayout(LayoutKind.Sequential)]
+/// <remarks>
+/// Size is padded to 16 bytes so
+/// </remarks>
+[StructLayout(LayoutKind.Sequential, Size = 16)]
 internal struct PvsData
 {
     /// <summary>
@@ -146,17 +152,44 @@ internal struct PvsData
 /// <summary>
 /// Specialized struct with the same size as <see cref="PvsData"/> that is used to store metadata in the pinned PVsData array
 /// </summary>
-[StructLayout(LayoutKind.Sequential)]
+[StructLayout(LayoutKind.Sequential, Size = 16)]
 internal struct PvsMetadata
 {
     /// <summary>
     /// Tick at which this entity was last sent to a player.
     /// </summary>
     public NetEntity NetEntity;
+
     public GameTick LastModifiedTick;
     public ushort VisMask;
     public EntityLifeStage LifeStage;
-    public byte Unused;
+#if DEBUG
+    // This struct is padded to a size of 16 so it's aligned to cache boundaries nicely.
+    // We have this extra space that isn't being used,
+    // so I'm opting to use them to make debugging the free list easier.
+    // "Marker" overlaps with the field used by the free list (which occupies the unused memory of PvsMetadata).
+    // So we set it to a bogus value and BAM! Errors are obvious!
+    private byte Pad0;
+    public uint Marker;
+#endif
+}
+
+[StructLayout(LayoutKind.Sequential, Size = 16)]
+internal struct PvsMetadataFreeLink
+{
+#if DEBUG
+    static unsafe PvsMetadataFreeLink()
+    {
+        DebugTools.Assert(sizeof(PvsMetadataFreeLink) == sizeof(PvsMetadata));
+    }
+#endif
+
+    public int Pad0;
+    public int Pad1;
+    public int Pad2;
+    // We offset the NextFree to be at the end of the struct.
+    // This is so that it overlaps with the debug Marker field of PvsMetadata instead of real data.
+    public PvsIndex NextFree;
 }
 
 /// <summary>