--- a/Robust.Server/GameStates/PvsSystem.Ack.cs
+++ b/Robust.Server/GameStates/PvsSystem.Ack.cs
@@ -85,10 +85,10 @@ internal sealed partial class PvsSystem
     /// <summary>
     ///     Process a given client's queued ack.
     /// </summary>
-    private void ProcessQueuedAck(PvsSession session)
+    private unsafe void ProcessQueuedAck(PvsSession session)
     {
         var ackedTick = session.LastReceivedAck;
-        List<PvsData>? ackedEnts;
+        List<IntPtr>? ackedEnts;
 
         if (session.Overflow != null && session.Overflow.Value.Tick <= ackedTick)
         {
@@ -108,12 +108,15 @@ internal sealed partial class PvsSystem
         else if (!session.PreviouslySent.TryGetValue(ackedTick, out ackedEnts))
             return;
 
-        foreach (var data in CollectionsMarshal.AsSpan(ackedEnts))
+        foreach (ref var intPtr in CollectionsMarshal.AsSpan(ackedEnts))
         {
-            data.EntityLastAcked = ackedTick;
+            ValidatePtr(intPtr);
+#if DEBUG
+            var data = *(PvsData*)intPtr;
+            DebugTools.AssertNotEqual(data.LastSeen, GameTick.Zero);
             DebugTools.Assert(data.LastSeen >= ackedTick); // LastSent may equal ackedTick if the packet was sent reliably.
-            DebugTools.Assert(!session.Entities.TryGetValue(data.NetEntity, out var old)
-                              || ReferenceEquals(data, old));
+#endif
+            ((PvsData*)intPtr)->EntityLastAcked = ackedTick;
         }
 
         // The client acked a tick. If they requested a full state, this ack happened some time after that, so we can safely set this to false