--- a/Robust.Server/GameStates/PvsSystem.Pooling.cs
+++ b/Robust.Server/GameStates/PvsSystem.Pooling.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using Microsoft.Extensions.ObjectPool;
 using Robust.Shared.GameObjects;
@@ -15,8 +16,8 @@ internal sealed partial class PvsSystem
     /// </summary>
     private const int MaxVisPoolSize = 1024;
 
-    private readonly ObjectPool<List<PvsData>> _entDataListPool
-        = new DefaultObjectPool<List<PvsData>>(new ListPolicy<PvsData>(), MaxVisPoolSize);
+    private readonly ObjectPool<List<IntPtr>> _entDataListPool
+        = new DefaultObjectPool<List<IntPtr>>(new ListPolicy<IntPtr>(), MaxVisPoolSize);
 
     private readonly ObjectPool<HashSet<EntityUid>> _uidSetPool
         = new DefaultObjectPool<HashSet<EntityUid>>(new SetPolicy<EntityUid>(), MaxVisPoolSize);