@ElectroJr 2024-03-16
I don't fully understand how some of the new memory stuff works, but I trust PJB. Benchmark wise its about the same speed as my pointer stuff. Benchmark results have changed relative to the original PR, but that's probably mainly because ~2 months have gone by and the benchmarks depend on Content maps & code.


Master / Dictionaries :
| Method     | PlayerCount | Mean     | Error     | StdDev    |
|----------- |------------ |---------:|----------:|----------:|
| Static   | 80          | 3.393 ms | 0.0172 ms | 0.0160 ms |
| Teleporting | 80          | 5.096 ms | 0.0275 ms | 0.0257 ms |

Old:
| Method     | PlayerCount | Mean     | Error     | StdDev    |
|----------- |------------ |---------:|----------:|----------:|
| Static   | 80          | 1.283 ms | 0.0058 ms | 0.0051 ms |
| Teleporting | 80          | 4.327 ms | 0.0162 ms | 0.0126 ms |


New memory managment:
| Method     | PlayerCount | Mean     | Error     | StdDev    |
|----------- |------------ |---------:|----------:|----------:|
| Static   | 80          | 1.313 ms | 0.0053 ms | 0.0044 ms |
| Teleporting | 80          | 4.331 ms | 0.0397 ms | 0.0372 ms |